% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF.R
\name{DF}
\alias{DF}
\title{Calculate Degrees of Freedom for Ridge Regression}
\usage{
DF(lambda, X)
}
\arguments{
\item{lambda}{A numeric scalar for the penalty parameter.}

\item{X}{A numeric matrix of predictors (training data).}
}
\value{
A single numeric value: the degrees of freedom.
}
\description{
This function computes the effective degrees of freedom (DF) for a
Ridge regression model given a penalty parameter lambda.
}
