% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLikelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Calculate the Log-Likelihood for a Linear Model with Intercept of Zero}
\usage{
LogLikelihood(theta, sigma, X, y)
}
\arguments{
\item{theta}{A numeric vector of model parameters (β).}

\item{sigma}{A single numeric value for the error standard deviation.}

\item{X}{A numeric matrix of predictors (training data).}

\item{y}{A numeric vector of the response (training data).}
}
\value{
A single numeric value: the total log-likelihood.
}
\description{
This function computes the total log-likelihood for a linear model
y = theta*X + epsilon, assuming epsilon are i.i.d. Normal(0, sigma^2).
}
