% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge.R
\name{Ridge}
\alias{Ridge}
\title{Calculate the Ridge Regression Cost Function}
\usage{
Ridge(theta, sigma, X, y, lambda)
}
\arguments{
\item{theta}{A numeric vector of model parameters (β).}

\item{sigma}{A single numeric value for the error standard deviation.}

\item{X}{A numeric matrix of predictors (training data).}

\item{y}{A numeric vector of the response (training data).}

\item{lambda}{A single numeric value for the penalty parameter.}
}
\value{
A single numeric value: the total cost (Minus Log-Likelihood + Ridge penalty).
}
\description{
This function computes the penalized minus log-likelihood for a
Ridge regression model. This is the cost function to be *minimized*.
}
